requirements.txt
requests==2.32.5
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
